Fix FutureWarning related to chain assignment using an inplace method
FutureWarning: A value is trying to be set on a copy of a DataFrame or
Series through chained assignment using an inplace method. The behavior
will change in pandas 3.0. This inplace method will never work because
the intermediate object on which we are setting values always behaves as
a copy.